# Faster local dofs