This branch is based on https://github.com/ONSAS/ONSAS.jl/pull/388,  but is independent.

---

Before:

```julia
julia> @benchmark local_dofs($tetra)
BenchmarkTools.Trial: 10000 samples with 84 evaluations.
 Range (min … max):  813.988 ns … 31.485 μs  ┊ GC (min … max):  0.00% … 96.09%
 Time  (median):     831.345 ns              ┊ GC (median):     0.00%
 Time  (mean ± σ):   938.150 ns ±  1.722 μs  ┊ GC (mean ± σ):  10.43% ±  5.51%

      ▂█▆▄▅                                                     
  ▁▂▄▆██████▆▆▄▄▄▃▃▃▂▃▂▂▂▂▂▂▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  814 ns          Histogram: frequency by time          943 ns <

 Memory estimate: 1.77 KiB, allocs estimate: 19.
```


After:

```julia
julia> @benchmark local_dofs($tetra)
BenchmarkTools.Trial: 10000 samples with 864 evaluations.
 Range (min … max):  137.731 ns …   3.026 μs  ┊ GC (min … max):  0.00% … 94.03%
 Time  (median):     140.191 ns               ┊ GC (median):     0.00%
 Time  (mean ± σ):   168.730 ns ± 270.235 ns  ┊ GC (mean ± σ):  16.04% ±  9.42%

  █                                                              
  █▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂ ▂
  138 ns           Histogram: frequency by time         2.79 μs <

 Memory estimate: 544 bytes, allocs estimate: 4.
```