@mforets 2023-07-01
Interestingly, the error was shown with the newly added test (cantilever). As it turns out, traversal order matters, I left a comment in the modified method.